Creates pod.yaml
This describes a single instance of a container, including the container image, command to run, environment variables, and other settings.

The apiVersion specifies the version of the Kubernetes API that the YAML file uses.
The kind specifies that this is a Pod object.
The metadata section specifies the name of the pod.
The spec section contains the specification of the pod, including the container(s) to be run.
In this case, there is one container, named expert-container, which uses the my-expert-image Docker image with the latest tag.
The command and args fields are used to specify the command that the container should run.
In this case, the command is a shell command that prints a message every 10 seconds, demonstrating that the pod is running.

This is a simple example, but you can add more complexity and customization as needed for your expert pod. For example, you might want to specify environment variables, mount volumes, or use a specific networking configuration.