Create PersistentVolume.yaml
In this example, we are creating a PersistentVolume with the name "pv-example" and a capacity of 10Gi. The access mode is set to ReadWriteOnce, which means that it can be mounted by a single node for both reading and writing. The persistentVolumeReclaimPolicy is set to Retain, which means that the data will not be deleted when the volume is released.

The storageClassName is set to "slow", which is a class of storage with specific performance and availability characteristics. The mountOptions are set to configure the NFS server connection with specific parameters, including the NFS protocol version, read and write buffer sizes, and the type of connection.

The nfs section specifies the NFS server details, including the server hostname and the mount path. The nodeAffinity section is used to specify the nodes that this PersistentVolume should be available on, using a nodeSelectorTerm that matches on the kubernetes.io/hostname label.

This is just an example, and the specific configuration of a PersistentVolume will depend on the specific storage requirements of the application being deployed.