Create PersistentVolumeClaim.yaml
In this example, the PersistentVolumeClaim is named my-pvc and is created in the my-namespace namespace. Here's a breakdown of the different fields:

accessModes: The access mode of the volume. In this case, the volume can be mounted in read-write mode by a single node at a time (ReadWriteOnce).
resources: The amount of storage requested by the PVC. In this case, the volume will be at least 50Gi in size.
storageClassName: The name of the StorageClass that the PVC should use. In this case, it is set to slow.
volumeMode: The mode in which the volume should be formatted. In this case, it is set to Filesystem.
selector: The labels used to match a specific Persistent Volume. In this case, it is matching app: my-app.
annotations: Additional metadata for the PVC. In this case, it specifies the storage class and the storage provisioner.
Note that the specific values used in the YAML file will depend on your specific needs and your Kubernetes cluster configuration.