Create StatefulSet.yaml
This StatefulSet.yaml file specifies a StatefulSet named my-statefulset with three replicas. Each replica will use a container named my-container running the my-image:latest Docker image, with port 80 exposed. The container will mount a persistent volume named my-volume to the path /data.

The StatefulSet will use a service named my-service, and each pod will be labeled with app: my-app. Additionally, the StatefulSet will create a volume claim template for a persistent volume claim named my-pvc with an annotation to use a specific storage class. The persistent volume claim will request 1Gi of storage with ReadWriteOnce access mode.

This is just an example, and the specifics of the file would vary depending on the requirements of the application being deployed.