Create DaemonSet.yaml
In this example, we're creating a DaemonSet called "my-app" with a single container called "my-container". We're specifying the container image to use, its resource requirements, environment variables, and readiness and liveness probes. We're also mounting a volume from a ConfigMap to the container's file system. Finally, we're using a RollingUpdate strategy for updates to the DaemonSet.

Note that this is just one example of what an advanced DaemonSet.yaml file could look like. The specific contents will depend on the needs of your application and infrastructure.